feat: renegotiation

+++ app/src/main/kotlin/com/android/webrtc/example/view/MainActivity.kt
@@ -1,8 +1,12 @@
 package com.android.webrtc.example.view
 
 import android.os.Bundle
+import androidx.activity.OnBackPressedCallback
 import androidx.appcompat.app.AppCompatActivity
+import androidx.navigation.findNavController
+import com.android.webrtc.example.R
 import com.android.webrtc.example.databinding.ActivityMainBinding
+import kotlin.concurrent.fixedRateTimer
 
 class MainActivity : AppCompatActivity() {
 
@@ -19,7 +23,12 @@ class MainActivity : AppCompatActivity() {
     // !!! Don't do it in prod, it's workaround to avoid memory leaks
     // We don't care for any cache, unfinished jobs, etc...
     override fun onBackPressed() {
-        android.os.Process.killProcess(android.os.Process.myPid())
+        val fragmentsCount =
+            supportFragmentManager.fragments.firstOrNull()?.childFragmentManager?.backStackEntryCount
+        if (fragmentsCount == 0) {
+            // exit app
+            android.os.Process.killProcess(android.os.Process.myPid())
+        }
         super.onBackPressed()
     }
 }
